+++ fj-core/src/main/java/org/fugerit/java/core/db/daogen/BasicHelper.java
@@ -1,6 +1,7 @@
 package org.fugerit.java.core.db.daogen;
 
 import java.io.Serializable;
+import java.util.function.BooleanSupplier;
 
 public class BasicHelper implements Serializable {
 
@@ -9,10 +10,14 @@ public class BasicHelper implements Serializable {
 	 */
 	private static final long serialVersionUID = 7545568256714803873L;
 
-	protected static final boolean UNSOPPORTED_OPERATION = true;
-	
+	private static final boolean UNSUPPORTED_OPERATION = true;
+
 	public static void throwUnsupported( String message ) {
-		if ( UNSOPPORTED_OPERATION ) {
+		throwUnsupported( message, () -> UNSUPPORTED_OPERATION );
+	}
+
+	public static void throwUnsupported(String message, BooleanSupplier isUnsupported) {
+		if ( isUnsupported.getAsBoolean() ) {
 			throw new UnsupportedOperationException( message );
 		}
 	}