+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- BasicWrapperNG with no serialization
+
 ## [8.5.1] - 2024-03-08
 
 ### Added